+++ model/ReadingTimerUnitType.ts
@@ -1,0 +1,3 @@
+export enum ReadingTimerUnitType {
+  Pages, Entries
+}