+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### Version 1.1
+- Added reading timer tracking elapsed time and reading speed
 - Fixed bug where text selection is not cleared when switching pages
 - Text hooker WebSocket URL can now be changed in settings